--- a/package.json
+++ b/package.json
@@ -24,7 +24,7 @@
     "babel-plugin-transform-react-jsx": "^6.8.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
-    "buble": "^0.15.2",
+    "buble": "^0.17.3",
     "eslint-config-xo-swizz": "^0.11.0",
     "rimraf": "^2.6.1",
     "rollup": "^0.41.3",