--- a/perf/.gitignore
+++ b/perf/.gitignore
@@ -1,0 +1,2 @@
+/snnabdom-new.js
+/snnabdom-base.js