--- a/perf/build-branches.js
+++ b/perf/build-branches.js
@@ -1,11 +1,13 @@
-const execSync = require('child_process').execSync
-const fs = require('fs');
-const simpleGit = require('simple-git')();
+/* eslint unicorn/no-process-exit: 0 */
 
-let baseBranch =  process.argv[2] || 'develop'
+const execSync = require('child_process').execSync
+const fs = require('fs')
+const simpleGit = require('simple-git')()
+
+const baseBranch = process.argv[2] || 'develop'
 
 function buildSnabbdom(suffix) {
-  execSync('npm run build', {stdio: [0, 1, 2]});
+  execSync('npm run build', { stdio: [0, 1, 2] })
   fs.writeFileSync(`perf/snnabdom-${suffix}.js`, fs.readFileSync('dist/index.js', 'utf-8'), 'utf-8')
 }
 
@@ -21,18 +23,18 @@ simpleGit.status((err, status) => {
     } else {
       console.log('Building ' + currentBranch + ' branch')
       buildSnabbdom('new')
-      simpleGit.checkout(baseBranch, (err, data) => {
+      simpleGit.checkout(baseBranch, (err) => {
         if (err) {
           console.error('Error checkout ' + baseBranch)
         } else {
           console.log('Building ' + baseBranch + ' branch')
-          buildSnabbdom('base')          
+          buildSnabbdom('base')
         }
         simpleGit.checkout(currentBranch)
         if (err) {
           process.exit(1)
         }
       })
-    }  
+    }
   }
 })