--- a/perf/modules-object-props.js
+++ b/perf/modules-object-props.js
@@ -1,33 +1,35 @@
-var Benchmark = require('benchmark');
-var h = require('snabbdom/h').default;
-var baseCreateElement = require('./snnabdom-base').createElement
-var newCreateElement = require('./snnabdom-new').createElement
+/* eslint import/no-unresolved: 0 */
 
-function objectPropsHyper () {
+const Benchmark = require('benchmark')
+const h = require('snabbdom/h').default
+const baseCreateElement = require('./snnabdom-base').createElement
+const newCreateElement = require('./snnabdom-new').createElement
+
+function objectPropsHyper() {
   return h('div', { style: { color: 'red', fontWeight: 'bold' }, dataset: { color: 'red', hidden: true } }, [
-    h('span', {attrs: {href: '/foo'}}, 'Hello'),
-    h('span', {dataset: {action: 'reset'}}, 'World')
+    h('span', { attrs: { href: '/foo' } }, 'Hello'),
+    h('span', { dataset: { action: 'reset' } }, 'World')
   ])
 }
 
-function objectProps (createElement) {
-  return createElement('div', { style: { color: 'red', fontWeight: 'bold' }, data: { color: 'red', hidden: true }}, [
-    createElement('span', {attrs: {href: '/foo'}}, 'Hello'),
-    createElement('span', {data: {action: 'reset'}}, 'World')
+function objectProps(createElement) {
+  return createElement('div', { style: { color: 'red', fontWeight: 'bold' }, data: { color: 'red', hidden: true } }, [
+    createElement('span', { attrs: { href: '/foo' } }, 'Hello'),
+    createElement('span', { data: { action: 'reset' } }, 'World')
   ])
 }
 
-var suite = new Benchmark.Suite('module object props');
+const suite = new Benchmark.Suite('module object props')
 
 // add tests
-suite.add('hyperscript', function() {
+suite.add('hyperscript', () => {
   objectPropsHyper()
-})
-.add('jsx-base', function() {
+}).
+add('jsx-base', () => {
   objectProps(baseCreateElement)
-})
-.add('jsx-new', function() {
+}).
+add('jsx-new', () => {
   objectProps(newCreateElement)
-});
+})
 
 module.exports = suite