--- a/perf/simple-props.js
+++ b/perf/simple-props.js
@@ -1,33 +1,35 @@
-var Benchmark = require('benchmark');
-var h = require('snabbdom/h').default;
-var baseCreateElement = require('./snnabdom-base').createElement
-var newCreateElement = require('./snnabdom-new').createElement
+/* eslint import/no-unresolved: 0 */
 
-function simplePropsHyper () {
-  return h('div', {props: {href: 'xxxxx', target: 'route'}}, [
-    h('span', {props: {name: 'x', value: 1}}, 'Hello'),
-    h('span', {props: {name: 'y', value: 2}}, 'World')
+const Benchmark = require('benchmark')
+const h = require('snabbdom/h').default
+const baseCreateElement = require('./snnabdom-base').createElement
+const newCreateElement = require('./snnabdom-new').createElement
+
+function simplePropsHyper() {
+  return h('div', { props: { href: 'xxxxx', target: 'route' } }, [
+    h('span', { props: { name: 'x', value: 1 } }, 'Hello'),
+    h('span', { props: { name: 'y', value: 2 } }, 'World')
   ])
 }
 
-function simpleProps (createElement) {
-  return createElement('div', {href: 'xxxxx', target: 'route'}, [
-    createElement('span', {name: 'x', value: 1}, 'Hello'),
-    createElement('span', {name: 'y', value: 2}, 'World')
+function simpleProps(createElement) {
+  return createElement('div', { href: 'xxxxx', target: 'route' }, [
+    createElement('span', { name: 'x', value: 1 }, 'Hello'),
+    createElement('span', { name: 'y', value: 2 }, 'World')
   ])
 }
 
-var suite = new Benchmark.Suite('simple props');
+const suite = new Benchmark.Suite('simple props')
 
 // add tests
-suite.add('hyperscript', function() {
+suite.add('hyperscript', () => {
   simplePropsHyper()
-})
-.add('jsx-base', function() {
+}).
+add('jsx-base', () => {
   simpleProps(baseCreateElement)
-})
-.add('jsx-new', function() {
+}).
+add('jsx-new', () => {
   simpleProps(newCreateElement)
-});
+})
 
 module.exports = suite