--- a/test/snabbdom-specs/vnode-class-names/actual.js
+++ b/test/snabbdom-specs/vnode-class-names/actual.js
@@ -2,5 +2,5 @@
 import { isVisible, isEnabled } from './neutral'
 
 export default (createElement) => {
-  return createElement('div', { 'class-visible': isVisible, 'class-enabled': isEnabled })
+  return createElement('div', { 'class-visible': isVisible, 'class-enabled': isEnabled, 'class-alert-danger': true })
 }