--- a/test/snabbdom-specs/vnode-class-names/expected.js
+++ b/test/snabbdom-specs/vnode-class-names/expected.js
@@ -3,6 +3,6 @@ import { isVisible, isEnabled } from './neutral'
 
 export default (h) => {
   return h('div', {
-    class: { visible: isVisible, enabled: isEnabled }
+    class: { visible: isVisible, enabled: isEnabled, 'alert-danger': true }
   }, [])
 }