--- a/test/snabbdom-specs/vnode-style/actual.js
+++ b/test/snabbdom-specs/vnode-style/actual.js
@@ -1,4 +1,4 @@
 
 export default (createElement) => {
-  return createElement('div', { 'style-color': 'red' })
+  return createElement('div', { 'style-color': 'red', 'style-background-color': 'blue' })
 }