--- a/test/snabbdom-specs/vnode-style/expected.js
+++ b/test/snabbdom-specs/vnode-style/expected.js
@@ -1,6 +1,6 @@
 
 export default (h) => {
   return h('div', {
-    style: { color: 'red' }
+    style: { color: 'red', 'background-color': 'blue' }
   }, [])
 }