… dist/index.es6.js
… dist/index.js
… yarn.lock