Add scripts that compares performance of the current branch checked in the local repository and a base one (develop).

Steps:
 * builds current branch
 * checkout develop
 * builds develop branch
 * checkout current branch
 * run perf suites based on benchmark.js

Add two npm scripts: 
  * perf:all -> build the two branches and run perf suite
  * perf -> run perf suite

This PR contains changes from #27 

Here is the output of perf suite comparing #27 and develop:

Benchmarking no props
hyperscript x 13,230,387 ops/sec ±5.07% (91 runs sampled)
jsx-base x 699,560 ops/sec ±2.24% (98 runs sampled)
jsx-new x 676,120 ops/sec ±2.98% (93 runs sampled)
no props - fastest is hyperscript   fastest JSX is  [ 'jsx-base' ]

Benchmarking simple props
hyperscript x 13,216,166 ops/sec ±0.73% (98 runs sampled)
jsx-base x 41,596 ops/sec ±0.76% (92 runs sampled)
jsx-new x 45,047 ops/sec ±0.57% (98 runs sampled)
simple props - fastest is hyperscript   fastest JSX is  [ 'jsx-new' ]

Benchmarking module object props
hyperscript x 9,655,921 ops/sec ±0.62% (101 runs sampled)
jsx-base x 41,222 ops/sec ±2.70% (97 runs sampled)
jsx-new x 34,262 ops/sec ±4.33% (82 runs sampled)
module object props - fastest is hyperscript   fastest JSX is  [ 'jsx-base' ]

Benchmarking modules attribute props
hyperscript x 7,928,043 ops/sec ±5.80% (81 runs sampled)
jsx-base x 28,325 ops/sec ±2.04% (89 runs sampled)
jsx-new x 29,938 ops/sec ±1.11% (102 runs sampled)
modules attribute props - fastest is hyperscript   fastest JSX is  [ 'jsx-new' ]

Benchmarking real world form
jsx-base x 2,844 ops/sec ±1.19% (99 runs sampled)
jsx-new x 3,024 ops/sec ±0.66% (101 runs sampled)
real world form - fastest is jsx-new   fastest JSX is  [ 'jsx-new' ]

All benchmarks complete