@blikblum 2018-05-18
For the record this is the results of my changes, that were published as [snabbdom-pragma-lite](https://www.npmjs.com/package/snabbdom-pragma-lite):

Benchmarking no props
hyperscript x 12,807,054 ops/sec ±2.19% (88 runs sampled)
jsx-base x 740,841 ops/sec ±0.40% (104 runs sampled)
jsx-new x 3,146,969 ops/sec ±0.56% (106 runs sampled)
no props - fastest is hyperscript   fastest JSX is  [ 'jsx-new' ]

Benchmarking simple props
hyperscript x 12,201,875 ops/sec ±0.56% (107 runs sampled)
jsx-base x 42,757 ops/sec ±0.69% (107 runs sampled)
jsx-new x 1,201,914 ops/sec ±0.47% (105 runs sampled)
simple props - fastest is hyperscript   fastest JSX is  [ 'jsx-new' ]

Benchmarking module object props
hyperscript x 9,248,187 ops/sec ±0.64% (107 runs sampled)
jsx-base x 43,626 ops/sec ±0.56% (107 runs sampled)
jsx-new x 596,333 ops/sec ±0.42% (101 runs sampled)
module object props - fastest is hyperscript   fastest JSX is  [ 'jsx-new' ]

Benchmarking modules attribute props
hyperscript x 9,070,325 ops/sec ±0.51% (109 runs sampled)
jsx-base x 30,560 ops/sec ±0.70% (107 runs sampled)
jsx-new x 320,799 ops/sec ±0.52% (102 runs sampled)
modules attribute props - fastest is hyperscript   fastest JSX is  [ 'jsx-new' ]

Benchmarking real world form
jsx-base x 2,921 ops/sec ±1.42% (103 runs sampled)
jsx-new x 34,647 ops/sec ±0.61% (102 runs sampled)
real world form - fastest is jsx-new   fastest JSX is  [ 'jsx-new' ]