@blikblum 2018-05-20
Not commiting the build at each PR generates smaller diffs making easier to see what really changed. 

This is the general rule in JS projects