@Swizz 2018-05-20
I want to request the build on a PR on master at least.

Master should be only a release using git flow
https://danielkummer.github.io/git-flow-cheatsheet/