@blikblum 2018-05-20
This PR is on develop. 
Should i commit the build files?
Revert the travis changes?

Any thing is good to me