@Swizz 2018-05-20
Go for Travis revert + build. 
I prefer to ensure the lib to be built on PR merge