@Swizz 2018-06-04
Npm published with [v2.7.0](https://github.com/Swizz/snabbdom-pragma/releases/tag/2.7.0)